Move to group all actions in one package
