+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/actions/SearchAction.java
@@ -1,9 +1,10 @@
-package org.lodder.subtools.multisubdownloader.gui.actions.search;
+package org.lodder.subtools.multisubdownloader.actions;
 
 import java.util.List;
 
 import org.lodder.subtools.multisubdownloader.Messages;
 import org.lodder.subtools.multisubdownloader.gui.actions.ActionException;
+import org.lodder.subtools.multisubdownloader.gui.actions.search.SearchSetupException;
 import org.lodder.subtools.multisubdownloader.gui.dialog.Cancelable;
 import org.lodder.subtools.multisubdownloader.gui.dialog.progress.StatusListener;
 import org.lodder.subtools.multisubdownloader.gui.dialog.progress.fileindexer.IndexingProgressListener;
